Rename object since otherwise api doc generatiomn fails with weird error
