+++ akka-http/src/main/scala/io/opencensus/scala/akka/http/TracingDirective.scala
@@ -7,7 +7,7 @@ import com.typesafe.scalalogging.LazyLogging
 import io.opencensus.scala.Tracing
 import io.opencensus.scala.akka.http.propagation.AkkaB3FormatPropagation
 import io.opencensus.scala.akka.http.trace.EndSpanResponse
-import io.opencensus.scala.akka.http.trace.HttpAttributes._
+import io.opencensus.scala.akka.http.trace.HttpExtractors._
 import io.opencensus.scala.http.{HttpAttributes, ServiceAttributes, ServiceData}
 import io.opencensus.scala.http.propagation.Propagation
 import io.opencensus.trace.{Span, Status}
